Moved ServerManager to lib.rs as PumpkinServer
